refactor: Posting -> Post로 네이밍 변경
