--- src/docs/asciidoc/posting/posting.adoc
@@ -1,44 +1,0 @@
-ifndef::snippets[]
-:snippets: ./build/generated-snippets
-endif::[]
-
-= POSTING API 문서
-:icons: font
-:source-highlighter: highlight.js
-:toc: left
-:toclevels: 1
-:sectlinks:
-
-== API 목록
-
-link:../keeper.html[API 목록으로 돌아가기]
-
-== *게시글 생성하기*
-
-=== 요청
-
-==== Request
-
-include::{snippets}/create-post/http-request.adoc[]
-
-==== Request Cookies
-
-include::{snippets}/create-post/request-cookies.adoc[]
-
-==== Query Parameters
-
-include::{snippets}/create-post/query-parameters.adoc[]
-
-==== Request Parts
-
-include::{snippets}/create-post/request-parts.adoc[]
-
-=== 응답
-
-==== Response
-
-include::{snippets}/create-post/http-response.adoc[]
-
-==== Response Headers
-
-include::{snippets}/create-post/response-headers.adoc[]
