+++ src/main/java/com/keeper/homepage/domain/member/entity/post/MemberHasPostingPK.java
@@ -1,0 +1,11 @@
+package com.keeper.homepage.domain.member.entity.post;
+
+import java.io.Serializable;
+import lombok.Getter;
+
+@Getter
+public class MemberHasPostingPK implements Serializable {
+
+  private Long member;
+  private Long post;
+}
