--- src/main/java/com/keeper/homepage/domain/member/entity/posting/MemberHasPostingLike.java
@@ -1,45 +1,0 @@
-package com.keeper.homepage.domain.member.entity.posting;
-
-import static jakarta.persistence.FetchType.LAZY;
-import static lombok.AccessLevel.PROTECTED;
-
-import com.keeper.homepage.domain.member.entity.Member;
-import com.keeper.homepage.domain.posting.entity.Posting;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.Id;
-import jakarta.persistence.IdClass;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.io.Serializable;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@EqualsAndHashCode(of = {"member", "posting"})
-@NoArgsConstructor(access = PROTECTED)
-@IdClass(MemberHasPostingPK.class)
-@Table(name = "member_has_posting_like")
-public class MemberHasPostingLike {
-
-  @Id
-  @ManyToOne(fetch = LAZY)
-  @JoinColumn(name = "member_id", nullable = false)
-  private Member member;
-
-  @Id
-  @ManyToOne(fetch = LAZY)
-  @JoinColumn(name = "posting_id", nullable = false)
-  private Posting posting;
-
-  @Builder
-  private MemberHasPostingLike(Member member, Posting posting) {
-    this.member = member;
-    this.posting = posting;
-  }
-}
