+++ src/main/java/com/keeper/homepage/domain/post/dao/PostRepository.java
@@ -1,0 +1,8 @@
+package com.keeper.homepage.domain.post.dao;
+
+import com.keeper.homepage.domain.post.entity.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post, Long> {
+
+}
