+++ src/main/java/com/keeper/homepage/domain/post/dao/category/CategoryRepository.java
@@ -1,0 +1,8 @@
+package com.keeper.homepage.domain.post.dao.category;
+
+import com.keeper.homepage.domain.post.entity.category.Category;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+
+}
