--- src/main/java/com/keeper/homepage/domain/posting/dao/PostingRepository.java
@@ -1,8 +1,0 @@
-package com.keeper.homepage.domain.posting.dao;
-
-import com.keeper.homepage.domain.posting.entity.Posting;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PostingRepository extends JpaRepository<Posting, Long> {
-
-}
