+++ src/main/java/com/keeper/homepage/global/error/ErrorCode.java
@@ -24,7 +24,7 @@ public enum ErrorCode {
   TITLE_TYPE_NOT_FOUND("해당 타입의 타이틀을 찾을 수 없습니다.", HttpStatus.NOT_FOUND),
   // SEMINAR
   SEMINAR_TYPE_NOT_FOUND("존재하지 않는 세미나 타입입니다.", HttpStatus.NOT_FOUND),
-  // POSTING
+  // POST
   CATEGORY_NOT_FOUND("존재하지 않는 카테고리입니다.", HttpStatus.NOT_FOUND),
   ;
